requirements.txt
pytorch
numpy
matplotlib
segmentation-models-pytorch
albumentationsx 
scikit-learn
pandas
torch
tensorboard
